requirements.txt
flask>=3.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
PyJWT>=2.8
bcrypt>=4.0
python-dotenv>=1.0
cryptography>=41.0
requests>=2.31
gunicorn>=21.2
apscheduler>=3.10
croniter>=2.0
anthropic>=0.25
pgvector>=0.3
numpy>=1.26
PyYAML>=6.0
